move ok_expect tests
