+++ tests/ui/ok_expect.stderr
@@ -1,0 +1,32 @@
+error: called `ok().expect()` on a Result value. You can call `expect` directly on the `Result`
+  --> $DIR/ok_expect.rs:14:5
+   |
+14 |     res.ok().expect("disaster!");
+   |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+   |
+   = note: `-D ok-expect` implied by `-D warnings`
+
+error: called `ok().expect()` on a Result value. You can call `expect` directly on the `Result`
+  --> $DIR/ok_expect.rs:20:5
+   |
+20 |     res3.ok().expect("whoof");
+   |     ^^^^^^^^^^^^^^^^^^^^^^^^^
+
+error: called `ok().expect()` on a Result value. You can call `expect` directly on the `Result`
+  --> $DIR/ok_expect.rs:22:5
+   |
+22 |     res4.ok().expect("argh");
+   |     ^^^^^^^^^^^^^^^^^^^^^^^^
+
+error: called `ok().expect()` on a Result value. You can call `expect` directly on the `Result`
+  --> $DIR/ok_expect.rs:24:5
+   |
+24 |     res5.ok().expect("oops");
+   |     ^^^^^^^^^^^^^^^^^^^^^^^^
+
+error: called `ok().expect()` on a Result value. You can call `expect` directly on the `Result`
+  --> $DIR/ok_expect.rs:26:5
+   |
+26 |     res6.ok().expect("meh");
+   |     ^^^^^^^^^^^^^^^^^^^^^^^
+
